+++ babel.config.mjs
@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   presets: [
       ['@babel/preset-env', {
           targets: {
