+++ src/ProfileView.ts
@@ -1,17 +1,17 @@
 import { html, render } from 'lit-html'
 import { LiveStore } from 'rdflib'
-import './ChatWithMeElement'
-import './FriendListElement'
+import './ChatWithMe'
+import './FriendList'
 import { presentFriends } from './FriendsPresenter'
 import { presentProfile } from './presenter'
 import { presentCV } from './CVPresenter'
 import { presentSocial } from './SocialPresenter'
 import { presentStuff } from './StuffPresenter'
-import './ProfileCardElement'
-import './SocialCardElement'
-import './StuffCardElement'
-import './CVCardElement'
-import './QRCodeCardElement'
+import './ProfileCard'
+import './SocialCard'
+import './StuffCard'
+import './CVCard'
+import './QRCodeCard'
 import globalCssText from './styles/global.css'
 import profileCardCss from './styles/ProfileCard.css'
