+++ src/QRCodeCard.ts
